refactor(#367): crew.ts thin-wrapper — protocol→core, pane-helpers→workspaces (#376)
* refactor(#367): crew.ts thin-wrapper — protocol→core, pane-helpers→workspaces

Move orchestration logic out of packages/cli/src/commands/crew.ts so the
file becomes a thin wiring layer: parse args → call a package function →
format output. Extracted units are now unit-testable without spawning real
processes.

Extracted to @cockpit/core (packages/core/src/crew-protocol.ts):
  isTurnAccepted, buildCompletionProtocol, shellQuote, titleFor,
  isCrewTitle, nameFromTitle, nextAutoName, reapCrewChildren (crew-lifecycle.ts)

Extracted to @cockpit/workspaces (packages/workspaces/src/crew-pane.ts):
  getFreePort, listProjectCrews, findCrew, resolveCaptainWorkspace,
  sendFirstTurnWhenReady

Kept in cli (behavior-sensitive / cross-package orchestration):
  runCrewSpawn (240 LOC — couples cockpitdCall + agents + workspaces + shared)
  runCrewSend/Read/Close/List (use cockpitdCall, already unit-tested via mocks)
  Commander command definitions

New tests:
  packages/core/src/__tests__/crew-protocol.test.ts — pure protocol functions
    including snapshot guard on buildCompletionProtocol output (#278)
  packages/workspaces/src/__tests__/crew-pane.test.ts — sendFirstTurnWhenReady
    polling and retry logic with injected runtime mock

reapCrewChildren gains optional execFn parameter for test injection.
sendFirstTurnWhenReady accepts Pick<RuntimeDriver, ...> instead of full driver.

* fix(#367): update side.ts to import resolveCaptainWorkspace+sendFirstTurnWhenReady from @cockpit/workspaces

* fix(#367): add resolveCaptainWorkspace+sendFirstTurnWhenReady to @cockpit/workspaces mock in side.test.ts

* fix(#367): complete @cockpit/workspaces mock in side.test.ts — add listProjectCrews, findCrew, getFreePort

crew spawn regression test dynamically imports crew.ts which needs all three;
previous fix only added resolveCaptainWorkspace+sendFirstTurnWhenReady.
